Feature/new forms

+++ src/Controller/Admin/AdminController.php
@@ -6,32 +6,38 @@ namespace App\Controller\Admin;
 
 use App\Entity\Collection;
 use App\Entity\Item;
-use App\Entity\Medium;
 use App\Entity\Tag;
 use App\Entity\User;
 use App\Entity\Wish;
 use App\Entity\Wishlist;
+use App\Service\DatabaseDumper;
 use App\Service\DiskUsageCalculator;
+use Doctrine\ORM\NonUniqueResultException;
+use Doctrine\ORM\NoResultException;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\StreamedResponse;
+use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use ZipStream\Option\Archive;
+use ZipStream\ZipStream;
 
 /**
- * Class AdminController
- *
- * @package App\Controller
- *
- * @Route("/admin")
  * @IsGranted("ROLE_ADMIN")
  */
 class AdminController extends AbstractController
 {
     /**
-     * @Route("", name="app_admin_index", methods={"GET"})
+     * @Route({
+     *     "en": "/admin",
+     *     "fr": "/admin"
+     * }, name="app_admin_index", methods={"GET"})
      *
      * @return Response
+     * @throws NoResultException
+     * @throws NonUniqueResultException
      */
     public function index() : Response
     {
@@ -48,13 +54,22 @@ class AdminController extends AbstractController
                 'tags' => $em->getRepository(Tag::class)->countAll(),
                 'wishlists' => $em->getRepository(Wishlist::class)->countAll(),
                 'wishes' => $em->getRepository(Wish::class)->countAll(),
-            ]
+            ],
+            'symfonyVersion' => Kernel::VERSION,
+            'phpVersion' => phpversion(),
+            'isOpcacheAvailable' => function_exists('opcache_get_status') && opcache_get_status() && opcache_get_status()['opcache_enabled']
         ]);
     }
 
     /**
-     * @Route("/clean", name="app_admin_clean", methods={"GET"})
+     * @Route({
+     *     "en": "/admin/clean",
+     *     "fr": "/admin/nettoyer"
+     * }, name="app_admin_clean", methods={"GET"})
      *
+     * @param string $publicPath
+     * @param TranslatorInterface $translator
+     * @param DiskUsageCalculator $diskUsageCalculator
      * @return Response
      */
     public function clean(string $publicPath, TranslatorInterface $translator, DiskUsageCalculator $diskUsageCalculator) : Response
@@ -62,7 +77,7 @@ class AdminController extends AbstractController
         $em = $this->getDoctrine()->getManager();
 
         //Get all paths in database (image + image thumbnail)
-        $sql = "SELECT m.path as path, m.thumbnail_path as thumbnailPath FROM koi_medium m;";
+        $sql = "SELECT m.path as path, m.thumbnail_path as thumbnailPath FROM koi_image m;";
         $stmt = $em->getConnection()->prepare($sql);
         $stmt->execute();
 
@@ -76,7 +91,7 @@ class AdminController extends AbstractController
 
         //Get all paths on disk
         $rii = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($publicPath.'/uploads'));
-        $diskPaths = array();
+        $diskPaths = [];
         foreach ($rii as $file) {
             if (!$file->isDir() && $file->getFileName() !== '.gitkeep') {
                 $diskPaths[] = str_replace($publicPath. '/', '', $file->getPathname());
@@ -84,9 +99,11 @@ class AdminController extends AbstractController
         }
 
         //Compute the diff and delete the diff
-        $diff = array_diff($diskPaths, $dbPaths);
+        $diff = \array_diff($diskPaths, $dbPaths);
         foreach ($diff as $path) {
-            unlink($publicPath.'/'.$path);
+            if (file_exists($publicPath.'/'.$path)) {
+                unlink($publicPath.'/'.$path);
+            }
         }
 
         $this->addFlash('notice', $translator->trans('message.files_deleted', ['%count%' => \count($diff)]));
@@ -99,5 +116,50 @@ class AdminController extends AbstractController
         $this->getDoctrine()->getManager()->flush();
 
         return $this->redirectToRoute('app_admin_index');
+    }
+
+    /**
+     * @Route({
+     *     "en": "/admin/backup",
+     *     "fr": "/admin/sauvegarde"
+     * }, name="app_admin_backup", methods={"GET"})
+     *
+     * @param DatabaseDumper $databaseDumper
+     * @return StreamedResponse
+     */
+    public function backup(DatabaseDumper $databaseDumper) : StreamedResponse
+    {
+        $users = $this->getDoctrine()->getRepository(User::class)->findAll();
+
+        return new StreamedResponse(function () use ($databaseDumper, $users) {
+            $options = new Archive();
+            $options->setContentType('text/event-stream');
+            $options->setFlushOutput(true);
+            $options->setSendHttpHeaders(true);
+
+            $zipFilename = (new \DateTime())->format('Ymd') . '-koillection-backup.zip';
+            $zip = new ZipStream($zipFilename, $options);
+
+            foreach ($users as $user) {
+                $path = $this->getParameter('kernel.project_dir').'/public/uploads/'. $user->getId();
+
+                $files = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($path), \RecursiveIteratorIterator::LEAVES_ONLY);
+                foreach ($files as $name => $file) {
+                    if (!$file->isDir()) {
+                        $zip->addFileFromStream('/public/uploads/'. $user->getId() . '/' . $file->getFilename(), fopen($file->getRealPath(), 'r'));
+                    }
+                }
+            }
+
+            $fh = fopen('php://memory', 'r+');
+            foreach ($databaseDumper->dump() as $row) {
+                fwrite($fh, $row);
+            }
+            rewind($fh);
+            $zip->addFileFromStream((new \DateTime())->format('Ymd') . '-koillection-export.sql', $fh);
+            fclose($fh);
+
+            $zip->finish();
+        });
     }
 }
